add a new kaggle cli group
