+++ agent/function/tech_leader.py
@@ -1,7 +1,5 @@
 import ast
 
-import questionary
-
 from agent.hub.utils import match_plan
 from agent.integration import read_csv_file
 from agent.types import Plan
